Add summary data to json and now the workload, version name is not hardcoded
Add new funtion for ES indexing

Add es module to index in place of curl

Change the filename as well as now compatible with python3
